services/parsing_text_api/requirements.txt
fastapi
uvicorn
aiofiles
google-cloud-storage
google-cloud-pubsub
python-multipart
beautifulsoup4
python-dotenv
PyMuPDF
pdfplumber
pytesseract
trafilatura
pdf2image